feat: Add fetchSectionsByChklId API function
The `fetchSectionsByChklId` function in `api-call.js` has been added to handle the API call for retrieving sections by checklist ID. This function makes use of the `axios` library to send a GET request to the specified endpoint. It returns the response data if successful, or an error message if an error occurs during the request.

This addition enhances the functionality of the API module by providing a convenient way to fetch sections based on checklist ID.