feat: Add downloadStateAsJson function to Home component
The Home component in src/components/Home.jsx has been updated to include a new function called downloadStateAsJson. This function allows users to download the current state of the form data as a JSON file. When the function is called, it converts the form data to JSON, creates a Blob object with the JSON data, and generates a download link for the JSON file. After the file is downloaded, a success message is displayed. This addition enhances the functionality of the Home component by providing a convenient way for users to save and download their form data.